consolidate types for airquality section
