+++ src/components/AirQualitySection.tsx
@@ -12,6 +12,7 @@ import AirQualityParamBox from "./AirQualityParamBox";
 import { MODRawDeviceDataResponse } from "lib/quant";
 import determineSourceOfData from "lib/determineSourceOfData";
 import { format, parse } from "date-fns";
+import { AirQualityDevices, CommonDeviceType } from "types";
 
 type ParamAQIStandardMap = {
   O3: number;
@@ -20,20 +21,10 @@ type ParamAQIStandardMap = {
   NO2: number;
   PM1: null;
 };
-type airQualityDevices = {
-  site: string;
-  value: string;
-  latitude: number;
-  longitude: number;
-  sensorId: string;
-  location: string;
-  color: string;
-};
 type DataType =
   | RawDeviceAverageDataResponse[]
   | MODRawDeviceDataResponse[]
   | [string];
-type CommonDeviceType = RawDeviceAverageDataResponse & MODRawDeviceDataResponse;
 const paramAQIStandardMap: ParamAQIStandardMap = {
   O3: 70,
   "PM2.5": 35,
@@ -45,7 +36,7 @@ const paramAQIStandardMap: ParamAQIStandardMap = {
 function isDataTypeQuant(data: DataType): data is MODRawDeviceDataResponse[] {
   return "Model_PM_PM1" in data;
 }
-const AirQualitySection = ({ devices }: { devices: airQualityDevices[] }) => {
+const AirQualitySection = ({ devices }: { devices: AirQualityDevices[] }) => {
   const [selectedSensor, setSelectedSensor] = useState("MOD-PM-00404");
   const sanitizedValue = devices.length > 0 ? selectedSensor : "";
   const { data: deviceData = [], error: deviceDataError } = useSWR<DataType>(
